backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
celery
redis
yt-dlp
openai
whisper
transformers
scipy
torch
faiss-cpu
python-dotenv
websockets
cohere
